+++ frontend/src/v5/ui/controls/resizableTableContext/resizableTable/resizableTable.component.tsx
@@ -18,45 +18,7 @@
 import { Resizer } from '../resizer/resizer.component';
 import { useContext } from 'react';
 import { ResizableTableContext } from '../resizableTableContext';
-import { ResizersContainers, Table, ResizerLine } from './resizableTable.styles';
-
-const ResizerComponent = ({ name, width }) => {
-	const { setWidth, getWidth, setIsResizing, isResizing, setResizerName, resizerName, isHidden } = useContext(ResizableTableContext);
-	const currentWidth = getWidth(name);
-	const hidden = isHidden(name);
-
-	const onResizeStart = () => {
-		setIsResizing(true);
-		setResizerName(name);
-	};
-	const onResize = (offset) => setWidth(name, currentWidth + offset);
-	const onResizeEnd = () => {
-		setIsResizing(false);
-		setResizerName('');
-	};
-	const handleMouseOver = () => setResizerName(name);
-	const handleMouseOut = () => {
-		if (!isResizing) setResizerName('');
-	};
-
-	if (hidden) return null;
-
-	return (
-		<ResizerLine
-			$offset={width}
-			onMouseOver={handleMouseOver}
-			onMouseOut={handleMouseOut}
-			$isResizing={isResizing}
-			$highlight={resizerName === name}
-		>
-			<Resizer
-				onResizeStart={onResizeStart}
-				onResize={onResize}
-				onResizeEnd={onResizeEnd}
-			/>
-		</ResizerLine>
-	);
-};
+import { ResizersContainers, Table } from './resizableTable.styles';
 
 export const ResizableTable = ({ children }) => {
 	const { getElements } = useContext(ResizableTableContext);
@@ -66,8 +28,8 @@ export const ResizableTable = ({ children }) => {
 		<Table>
 			{children}
 			<ResizersContainers>
-				{resizableElements.map(({ name, width }) => (
-					<ResizerComponent name={name} width={width} key={name} />
+				{resizableElements.map(({ name }) => (
+					<Resizer name={name} key={name} />
 				))}
 			</ResizersContainers>
 		</Table>
