+++ frontend/src/v5/ui/controls/resizableTableContext/resizableTable/resizableTable.styles.ts
@@ -15,30 +15,12 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 export const Table = styled.div`
 	width: 100%;
 	overflow-x: scroll;
 	position: relative;
-`;
-
-export const ResizerLine = styled.div<{ $offset: number, $isResizing: boolean, $highlight: boolean }>`
-	position: relative;
-	z-index: 10;
-	width: 0;
-	height: 100%;
-	margin-left: ${({ $offset }) => $offset - 1}px;
-	border: dashed 1px transparent;
-	pointer-events: all;
-
-	${({ $highlight, theme }) => $highlight && css`
-		border-color: ${theme.palette.primary.main};
-	`}
-
-	${({ $isResizing }) => $isResizing && css`
-		border-style: solid;
-	`}
 `;
 
 export const ResizersContainers = styled.div`
